Setup new App.js file
